swap, swap/chain: move testbackend to mock package
